Written Integration Tests For ThreadOperations::WaitUntilTimeout - when call throws exception - when call times out - when call doesnt time out
