+++ ChickenControlSystem.sln
@@ -9,9 +9,9 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "HardwareCommunicationLayer"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "HardwareComs.ControlLine", "HardwareComs.ControlLine\HardwareComs.ControlLine.csproj", "{174A753F-0620-4237-BDFD-46E1D9475103}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "DoorModule", "DoorModule\DoorModule.csproj", "{82CF0491-91DF-4D32-B8B7-C9A472398AAB}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "HardwareAbstraction.DoorModule", "HardwareAbstraction.DoorModule\HardwareAbstraction.DoorModule.csproj", "{82CF0491-91DF-4D32-B8B7-C9A472398AAB}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "LightModule", "LightModule\LightModule.csproj", "{9B6246A6-A58D-4FC6-A9FF-2EDD51B8B37E}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "HardwareAbstraction.LightModule", "HardwareAbstraction.LightModule\HardwareAbstraction.LightModule.csproj", "{9B6246A6-A58D-4FC6-A9FF-2EDD51B8B37E}"
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "ControlSystem", "ControlSystem", "{AB571DB7-9CD8-4B28-B42A-0E266727EB4E}"
 EndProject
@@ -22,6 +22,8 @@ EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "HardwareComs.Tests.ControlLine.Integration", "HardwareComs.Tests.ControlLine.Integration\HardwareComs.Tests.ControlLine.Integration.csproj", "{C817DE50-7FB9-4494-A7DF-56B634C33B0B}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "HardwareComs.Tests.ControlLine.Unit", "HardwareComs.Tests.ControlLine.Unit\HardwareComs.Tests.ControlLine.Unit.csproj", "{71E53473-59D5-429B-92FD-50862AB3D2A5}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Tests", "Tests", "{4EB5A6ED-27FD-4121-A83A-885CBF4A3AD0}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -65,6 +67,7 @@ Global
 		{3205F8D9-F0DF-474C-BD24-3AB0DF5DEDC1} = {A67788A1-F279-4769-BA52-3A0B31CCD671}
 		{C817DE50-7FB9-4494-A7DF-56B634C33B0B} = {3205F8D9-F0DF-474C-BD24-3AB0DF5DEDC1}
 		{71E53473-59D5-429B-92FD-50862AB3D2A5} = {3205F8D9-F0DF-474C-BD24-3AB0DF5DEDC1}
+		{4EB5A6ED-27FD-4121-A83A-885CBF4A3AD0} = {73764B87-25AB-46B6-B9BD-62BDB8DFF29C}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {EC6C8F85-99C9-4B88-ADD1-00937B62E4F2}
