+++ HardwareAbstraction.DoorModule/HardwareAbstraction.DoorModule.csproj
@@ -2,6 +2,7 @@
 
   <PropertyGroup>
     <TargetFramework>netcoreapp3.1</TargetFramework>
+    <RootNamespace>DoorModule</RootNamespace>
   </PropertyGroup>
 
 </Project>
