+++ HardwareAbstraction.LightModule/HardwareAbstraction.LightModule.csproj
@@ -2,6 +2,7 @@
 
   <PropertyGroup>
     <TargetFramework>netcoreapp3.1</TargetFramework>
+    <RootNamespace>LightModule</RootNamespace>
   </PropertyGroup>
 
 </Project>
