+++ HardwareComs.Tests.ControlLine.Integration/BerkeleyRawSocketClientTests/BerkeleyRawSocketServer.cs
@@ -1,0 +1,6 @@
+﻿namespace ControlLineIntegrationTests.BerkeleyRawSocketClientTests
+{
+    public class BerkeleyRawSocketServer
+    {
+    }
+}
