+++ HardwareComs.Tests.ControlLine.Integration/ControlLineSocketsTests/SendOperation/SendOperationTests.cs
@@ -1,0 +1,13 @@
+﻿using ControlLine.Sockets;
+
+namespace ControlLineIntegrationTests.ControlLineSocketsTests.SendOperation
+{
+    public class SendOperationTests
+    {
+        protected ControlLineSockets Sut;
+
+        protected void Init()
+        {
+        }
+    }
+}
