+++ HardwareComs.Tests.ControlLine.Integration/ThreadOperationsTests/WaitUntilTimeout/Scenarios/CallThrowsException.cs
@@ -1,0 +1,45 @@
+﻿using System.Net.Sockets;
+using System.Threading;
+using NUnit.Framework;
+
+namespace ControlLineIntegrationTests.ThreadOperationsTests.WaitUntilTimeout.Scenarios
+{
+    [TestFixture(240)]
+    [TestFixture(200)]
+    [TestFixture(100)]
+    [TestFixture(10)]
+    [Description("Given ThreadOperations.WaitUntilTimeout Is Called, When Error Occurs In Call")]
+    public class CallThrowsException : WaitUntilTimeoutTests
+    {
+        private readonly int _timeoutPeriod;
+
+        public CallThrowsException(int timeoutPeriod)
+        {
+            _timeoutPeriod = timeoutPeriod;
+        }
+
+        [SetUp]
+        protected new void Init()
+        {
+            base.Init();
+        }
+
+        private void When()
+        {
+            Sut.WaitUntilTimeout(SutCall, Timeout);
+        }
+
+        private byte[] SutCall()
+        {
+            Thread.Sleep(_timeoutPeriod);
+            throw new SocketException();
+        }
+
+        [Test]
+        [Description("Then Error Occurs")]
+        public void ErrorTest()
+        {
+            Assert.Throws<SocketException>(When);
+        }
+    }
+}
